requirements.txt
openai
elasticsearch
pandas
python-dotenv